split in two classes as the two Rules don't work quite well together
