+++ integrations/src/test/java/io/jenkins/plugins/casc/GlobalLibrariesTest.java
@@ -1,10 +1,7 @@
 package io.jenkins.plugins.casc;
 
 import io.jenkins.plugins.casc.misc.ConfiguredWithCode;
-import io.jenkins.plugins.casc.misc.ConfiguredWithReadme;
 import io.jenkins.plugins.casc.misc.JenkinsConfiguredWithCodeRule;
-import io.jenkins.plugins.casc.misc.JenkinsConfiguredWithReadmeRule;
-import jenkins.plugins.git.GitSCMSource;
 import org.jenkinsci.plugins.github_branch_source.BranchDiscoveryTrait;
 import org.jenkinsci.plugins.github_branch_source.ForkPullRequestDiscoveryTrait;
 import org.jenkinsci.plugins.github_branch_source.ForkPullRequestDiscoveryTrait.TrustPermission;
@@ -28,17 +25,6 @@ public class GlobalLibrariesTest {
 
     @Rule
     public JenkinsConfiguredWithCodeRule j = new JenkinsConfiguredWithCodeRule();
-
-    @Test
-    @ConfiguredWithReadme("workflow-cps-global-lib/README.md")
-    public void configure_global_library() throws Exception {
-        assertEquals(1, GlobalLibraries.get().getLibraries().size());
-        final LibraryConfiguration library = GlobalLibraries.get().getLibraries().get(0);
-        assertEquals("awesome-lib", library.getName());
-        final SCMSourceRetriever retriever = (SCMSourceRetriever) library.getRetriever();
-        final GitSCMSource scm = (GitSCMSource) retriever.getScm();
-        assertEquals("https://github.com/jenkins-infra/pipeline-library.git", scm.getRemote());
-    }
 
     @Issue("JENKINS-57557")
     @Test
